fix: integrate fronting/faketls, fix build errors and tests
- Rewrite handler.go with proper MTPROTO protocol support
- Add domain fronting + FakeTLS integration for Telegram connection
- Add Domain field to proxy.Config
- Fix faketls ClientHello SNI extension layout
- Fix mtproto duplicate declarations (remove encrypt.go, message.go)
- Fix binary.LittleEndian.Int64 -> Uint64 in decrypt.go
- Fix rsa.GenerateKeyPair -> rsa.GenerateKey (API change)
- Add PKCS7 padding to Encrypt/Decrypt
- Fix unused imports in blocklist.go, socks5.go, handshake.go
- Fix fronting/http.go BodyLength field (not valid in http.Request)
- Add telegramDCs fallback addresses

+++ internal/mtproto/decrypt.go
@@ -18,8 +18,8 @@ func DecodeMessage(data []byte) (*Payload, error) {
 		return nil, ErrInvalidMessage
 	}
 
-	authKeyID := binary.LittleEndian.Int64(data[0:8])
-	msgID := binary.LittleEndian.Int64(data[8:16])
+	authKeyID := int64(binary.LittleEndian.Uint64(data[0:8]))
+	msgID := int64(binary.LittleEndian.Uint64(data[8:16]))
 	msgData := data[16:]
 
 	return &Payload{
